requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pypdf
google-generativeai
